add comments and seperate classes into their own files
